externalLambda/requirements.txt
openai
newspaper3k
awslambdaric